Fix payload vlaidations

+++ workflow-server/src/services/user.service.ts
@@ -1,20 +1,27 @@
-import { UserRepository } from "../repositories/index.js";
+import { UserRepository, BoardRepository } from "../repositories/index.js";
 import { Service, Inject } from "typedi";
 import { IUser } from "../types/database/index.js";
 import { Pagination } from "../types/utils.type.js";
-import { UserDTO, BoardDTO } from "../types/dto/index.js";
-import { BoardMapper, UserMapper } from "../mappers/index.js";
+import { UserDTO, BoardSimpleDTO } from "../types/dto/index.js";
+import { BoardSimpleViewMapper, UserMapper } from "../mappers/index.js";
+import { NotFoundError } from "routing-controllers";
+import { UpdateUserPayload } from "../types/request/user.type.js";
 
 @Service()
 export class UserService {
   userRepository: UserRepository;
+  boardRepository: BoardRepository;
 
-  constructor(@Inject() userRepository: UserRepository) {
+  constructor(@Inject() userRepository: UserRepository, @Inject() boardRepository: BoardRepository) {
     this.userRepository = userRepository;
+    this.boardRepository = boardRepository;
   }
 
   async getUser(userId: string): Promise<UserDTO> {
     const user = await this.userRepository.getById(userId);
+    if (!user) {
+      throw new NotFoundError("User was not found.");
+    }
     return UserMapper(user);
   }
 
@@ -37,17 +44,33 @@ export class UserService {
     };
   }
 
+  async updateUser(userId: string, userData: UpdateUserPayload): Promise<UserDTO> {
+    const user = await this.userRepository.getById(userId);
+    if (!user) {
+      throw new NotFoundError("User was not found.");
+    }
+    user.name = userData.name ?? user.name;
+    user.surname = userData.surname ?? user.surname;
+    user.email = userData.email ?? user.email;
+    await this.userRepository.save(user);
+    return UserMapper(user);
+  }
+
   async updateUserInfo(userId: string, userData: IUser): Promise<UserDTO> {
     const user = await this.userRepository.updateUser(userId, userData);
     return UserMapper(user);
   }
 
-  async getUserPinnedBoards(userId: string): Promise<BoardDTO[]> {
+  async getUserPinnedBoards(userId: string): Promise<BoardSimpleDTO[]> {
     const boards = await this.userRepository.getUserPinnedCollection(userId);
-    return boards.map(BoardMapper);
+    return boards.map(BoardSimpleViewMapper);
   }
 
   async togglePinBoard(userId: string, boardId: string): Promise<boolean> {
+    const board = await this.boardRepository.getById(boardId);
+    if (!board) {
+      throw new NotFoundError("Board does not exist");
+    }
     const boards = await this.userRepository.getUserPinnedCollection(userId);
     if (boards.find(({ _id }) => _id.equals(boardId))) {
       await this.userRepository.removeBoardFromPinnedCollection(userId, boardId);
